slack users endpoint / source from slack users endpoint
