+++ functions/slack-question.js
@@ -5,9 +5,6 @@ const formData = require('form-data')
 const Mailgun = require('mailgun.js')
 const getGravatar = require('gravatar')
 const { App, AwsLambdaReceiver } = require('@slack/bolt')
-const { createClient } = require('@supabase/supabase-js')
-
-const supabase = createClient(process.env.SUPABASE_URL, process.env.SUPABASE_API_KEY)
 
 const awsLambdaReceiver = new AwsLambdaReceiver({
     signingSecret: process.env.SLACK_SIGNING_SECRET,
@@ -439,95 +436,6 @@ app.action('publish-button', async ({ ack, client, body }) => {
         }
         client.chat.update(messageUpdate)
     }
-})
-
-app.shortcut('publish_thread', async ({ shortcut, ack, client, logger }) => {
-    try {
-        await ack()
-        const user = await client.users.info({ user: shortcut.user.id })
-        if (!user.user.is_admin) return
-        const { data, error } = await supabase
-            .from('Messages')
-            .select('slack_timestamp, slug')
-            .eq('slack_timestamp', shortcut.message.ts)
-
-        const view = {
-            trigger_id: shortcut.trigger_id,
-            view: {
-                private_metadata: JSON.stringify({
-                    ts: shortcut.message.ts,
-                    channel: shortcut.channel.id,
-                    user: shortcut.user.id,
-                }),
-                callback_id: 'submit-db-button',
-                type: 'modal',
-                title: {
-                    type: 'plain_text',
-                    text: 'PostHog Q&A',
-                },
-                submit: {
-                    type: 'plain_text',
-                    text: 'Publish',
-                    emoji: true,
-                },
-                close: {
-                    type: 'plain_text',
-                    text: 'Cancel',
-                    emoji: true,
-                },
-                blocks: [
-                    {
-                        type: 'input',
-                        optional: false,
-                        element: {
-                            type: 'plain_text_input',
-                            action_id: 'slug',
-                        },
-                        label: {
-                            type: 'plain_text',
-                            text: 'Slug',
-                            emoji: true,
-                        },
-                    },
-                ],
-            },
-        }
-        if (data.length > 0 && data[0].slug) {
-            view.view.blocks[0].element.initial_value = data[0].slug
-            view.view.blocks[0].optional = true
-            view.view.submit.text = 'Update'
-        }
-        await client.views.open(view)
-    } catch (error) {
-        logger.error(error)
-    }
-})
-
-app.view('submit-db-button', async ({ ack, view }) => {
-    await ack()
-    const {
-        private_metadata,
-        state: { values },
-    } = view
-    const body = {
-        slug: null,
-    }
-
-    const { ts, channel, user } = JSON.parse(private_metadata)
-
-    Object.keys(values).forEach((valueKey) => {
-        Object.keys(body).forEach((bodyKey) => {
-            if (values[valueKey][bodyKey]) {
-                body[bodyKey] = values[valueKey][bodyKey].value
-            }
-        })
-    })
-
-    body.slack_timestamp = ts
-    body.slack_channel = channel
-    body.slack_user = user
-
-    await supabase.from('Messages').upsert([body])
 })
 
 exports.handler = async (event, context, callback) => {
